+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+- added factories
+- code optimization
+- class models for added styles
+
 ## 0.0.1
 
 ## 0.0.2
