+++ lib/src/list_selection_widget/package.dart
@@ -1,13 +1,5 @@
 export 'dart:async';
 
 export 'package:flutter/material.dart';
-export 'package:list_selection_widget/list_selection_widget.dart';
-export 'package:list_selection_widget/src/list_selection_widget/widgets/cross_animation_widget.dart';
-export 'package:list_selection_widget/src/list_selection_widget/widgets/list_selection_widget_decoration.dart';
-export 'package:list_selection_widget/src/list_selection_widget/widgets/list_selection_widget_item_content.dart';
-export 'package:list_selection_widget/src/list_selection_widget/widgets/list_selection_widget_items_decoration.dart';
-export 'package:list_selection_widget/src/list_selection_widget/widgets/list_selection_widget_top_content.dart';
+export 'package:list_selection_widget/src/list_selection_widget/list_selection_widget.dart';
 export 'package:list_selection_widget/src/model/item_model.dart';
-export 'package:list_selection_widget/src/services/provider.dart';
-
-export 'widget.dart';
