+++ lib/src/list_selection_widget/widgets/list_selection_widget_top_content.dart
@@ -1,3 +1,4 @@
+import '../../services/provider.dart';
 import '../package.dart';
 
 class ListSelectionWidgetTitleContent extends StatefulWidget {
